Moving code blocks to new homes
Packaging termbox logic into termboxutil, and clustering like-natured code in main package to separate files
